Fixed render functions
