+++ Headers/config.h
@@ -2,6 +2,7 @@
 
 #include "include_all.h"
 
+// model type id
 #define MODEL_TYPE_TYPE GLuint
 #define MODEL_TYPE_MODELBASE 0
 #define MODEL_TYPE_MODEL 1
@@ -9,3 +10,13 @@
 #define MODEL_TYPE_BSPLINE 3
 #define MODEL_TYPE_DISCRETECURVE 4
 #define MODEL_TYPE_BEZIERSURFACE 5
+
+// model - modelBase adapters
+#define MODEL2MODELBASE ModelBaseParams{params.programID,params.programSelectedID,params.name,params.show,params.drawMode}
+#define BEZIERCURVE2MODELBASE ModelBaseParams{params.programID,params.programSelectedID,params.name,params.show,GL_LINE_STRIP}
+#define BSPLINE2MODELBASE ModelBaseParams{params.programID,params.programSelectedID,params.name,params.show,GL_LINE_STRIP}
+#define DISCRETECURVE2MODELBASE ModelBaseParams{params.programID,params.programSelectedID,params.name,params.show,GL_LINE_STRIP}
+#define BEZIERSURFACE2MODELBASE ModelBaseParams{params.programID,params.programSelectedID,params.name,params.show,GL_TRIANGLES}
+
+// for <math.h>
+#define _USE_MATH_DEFINES
