+++ Shaders/Modules/Color/Color_uniforms.glsl
@@ -1,4 +1,4 @@
 struct ColorUniforms{
 	vec3 color;
 };
-uniform ColorUniforms ColorData;
+uniform ColorUniforms colorData;
