+++ Sources/Models/Model.cpp
@@ -12,9 +12,6 @@ void Model::Render(RenderParams* p) {
 		return;
 	}
 
-	return;
-
-	/*
 	// Log::logToConsole("Render model: ", GetName());
 
 	MeshRenderParams mp {
@@ -62,7 +59,6 @@ void Model::Render(RenderParams* p) {
 			mesh->RenderSelection(&msp);
 		}
 	}
-	*/
 	
 }
 void Model::RenderSelection(RenderParams* p) {
