+++ imgui.ini
@@ -201,7 +201,7 @@ Collapsed=1
 
 [Window][Object editor]
 Pos=0,367
-Size=587,356
+Size=587,710
 
 [Window][Cubic B-Spline interpolation window]
 Pos=781,0
@@ -209,6 +209,6 @@ Size=460,319
 Collapsed=1
 
 [Window][Light options]
-Pos=10,22
-Size=502,301
+Pos=588,1
+Size=502,366
